+++ changelog.md
@@ -22,6 +22,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Add the experimental [`astar`](https://uiua.org/docs/astar) modifier, which performs the A* pathfinding algorithm
 - Add the experimental [`but ⫯`](https://uiua.org/docs/but) modifier, which is a complement to [`on ⟜`](https://uiua.org/docs/on) and [`by ⊸`](https://uiua.org/docs/by)
 - Non-scalar switch functions and [`repeat ⍥`](https://uiua.org/docs/repeat) now follow the same distribution and [`fix ¤`](https://uiua.org/docs/fix) rules as [`rows ≡`](https://uiua.org/docs/rows)
+- Add the `# Track caller!` semantic comment, which prevents stack traces from going below the function that contains it
 - Deprecate [`coordinate ⟔`](https://uiua.org/docs/coordinate) as it is seldom needed and easy to implement with other functions
   - It can be mostly replaced with `⊢⊚⌕`
 - Adjacent [`trace ⸮`](https://uiua.org/docs/trace)s now function as a single [`trace ⸮`](https://uiua.org/docs/trace) of more values